Fix bug with pending, unexpected closes
When going through preload.js, I noticed that the app properly handles unknown disconnects by removing them from the connections list, but if they're not yet connected (in pending state) it won't handle properly.

This update first checks if the closer is connected or awaiting approval. If connected, it will close the connection and update the status. Otherwise, it will simply remove them from the pending list and update the connections UI.

P.S. the reason we don't need this handling in onconnectionstatechange is because pending peers never connected to begin with! Therefore they don't have their own WebRTC connection yet and won't ever have this event listener.